format as operator
The line `"%d bottles of beer on the wall" format j - 100 should be(res0)` shows that `format` is both a member function of the type String and also an operator on the type String.  I wanted to call attention to that for those like me who are learning Scala through these exercises.